Table -- Font Awesome
Font awesome showing all icons except github forked. Suspected cause is css to postcss conversion,
with indented urls.

At the time of this commit, thisbounty.com was running a static site with no webpack or react. It
had a js include for font awesome, which requested a css file from the fa CDN. This css file
was manually downloaded to fa.css, converted to postcss, and bundled with site.

Modules exist for webpack. They are avoided due to having a lot of dependencies, like LESS. Tried
adding less, but did not compile