Table Frontend -- Mobile Elements
Added buttons for claim, flag, bookmark and delete

Be sure to check page with more than one bounty. Design
changes drastically with multiple rows. That's the reason for
the neutral colors.

Also, a possible feature is not all buttons will appear on every
bounty. Be sure to test design changes with that case as well.

Added button for new bounty. Plan is to allow users not logged
in to post a bounty, and then have a category for logged in,
paid, and guest bounty